reformat the api files
