+++ api/app.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const cors = require('cors');
-const { Sequelize } = require('sequelize');
+const sequelize = require('./config/database');
 
 const app = express();
 const port = 3000;
@@ -9,51 +9,6 @@ const port = 3000;
 app.use(cors());
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
-
-// Configuration de Sequelize pour PostgreSQL
-require('dotenv').config();
-const sequelize = new Sequelize(process.env.DB_NAME, process.env.DB_USER, process.env.DB_PASSWORD, {
-  host: 'localhost',
-  dialect: 'postgres'
-});
-
-sequelize.authenticate()
-  .then(() => {
-    console.log('Connection to the database has been established successfully.');
-  })
-  .catch(err => {
-    console.error('Unable to connect to the database:', err);
-  });
-
-// Définition d'un modèle exemple
-const User = sequelize.define('User', {
-  id: {
-    type: Sequelize.INTEGER,
-    autoIncrement: true,
-    primaryKey: true
-  },
-  username: {
-    type: Sequelize.STRING(50),
-    allowNull: false,
-    unique: true
-  },
-  email: {
-    type: Sequelize.STRING(100),
-    allowNull: false,
-    unique: true
-  },
-  password: {
-    type: Sequelize.STRING(255),
-    allowNull: false
-  },
-  createdAt: {
-    type: Sequelize.DATE,
-    defaultValue: Sequelize.NOW
-  }
-}, {
-  tableName: 'users',
-  timestamps: false // Si vous n'avez pas de colonnes `updated_at`, sinon mettre `true`
-});
 
 sequelize.sync({ force: true })
   .then(() => {
@@ -65,7 +20,4 @@ app.get('/', (req, res) => {
   res.json({ message: 'Hello from Express!' });
 });
 
-// Démarrage du serveur
-app.listen(port, () => {
-  console.log(`Server is running on http://localhost:${port}`);
-});
+module.exports = app;
