+++ api/config/database.js
@@ -1,0 +1,22 @@
+const { Sequelize } = require('sequelize');
+
+require('dotenv').config();
+const sequelize = new Sequelize(
+  process.env.DB_NAME, 
+  process.env.DB_USER, 
+  process.env.DB_PASSWORD, 
+  {
+    host: 'localhost',
+    dialect: 'postgres'
+  }
+);
+
+sequelize.authenticate()
+  .then(() => {
+    console.log('Connection to the database has been established successfully.');
+  })
+  .catch(err => {
+    console.error('Unable to connect to the database:', err);
+  });
+
+module.exports = sequelize;
