+++ api/server.js
@@ -1,0 +1,7 @@
+const app = require('./app');
+
+const PORT = 3000;
+
+app.listen(PORT, () => {
+  console.log(`Server is running on port ${PORT}`);
+});
